Moved reaction to its controller.
