+++ Group22Project/.classpath
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <classpath>
-	<classpathentry kind="src" path=""/>
+	<classpathentry kind="src" path="src"/>
 	<classpathentry kind="lib" path="C:/Users/ablha/Desktop/CPSC 219/Project/json-simple-1.1.jar">
 		<attributes>
 			<attribute name="module" value="true"/>
